# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"Campaign API","version":"0.1"},"paths":{"\/api\/check-eligibility":{"get":{"summary":"Customer eligible check","description":"Check for customer eligibility to redeem for voucher.","operationId":"checkEligibility","responses":{"200":{"description":"successful operation","content":{"application\/json":{"schema":{},"examples":{"eligible":{"summary":"Eligible","value":{"eligibility":true,"message":"Success."}},"eligible,locked":{"summary":"Eligible, voucher locked","value":{"eligibility":true,"message":"Voucher locked, proceed to validate."}},"not eligible,all redeemed":{"summary":"Not eligible, all redeemed","value":{"eligibility":false,"error_message":"All vouchers redeemed."}},"not eligible,all locked":{"summary":"Not eligible, all redeemed","value":{"eligibility":false,"error_message":"All vouchers have been claimed."}},"not eligible,redeemed":{"summary":"Not eligible, all redeemed","value":{"eligibility":false,"error_message":"Redeemed."}},"not eligible,transaction count":{"summary":"Not eligible, insufficient transaction count","value":{"eligibility":false,"error_message":"Less than 3 transactions in 30 days."}},"not eligible,transaction amount":{"summary":"Not eligible, insufficient transaction amount","value":{"eligibility":false,"error_message":"Total transactions less than $100."}}}}}}},"security":[{"bearer_token":[]}]}},"\/api\/validate-photo":{"post":{"summary":"Validate photo submission","description":"Uploaded photo will be validated throught image recognition API. Voucher code will be issued for success validation.","operationId":"validatePhoto","requestBody":{"required":true,"content":{"multipart\/form-data":{"schema":{"required":["photo"],"properties":{"photo":{"description":"photo to upload","type":"file"}},"type":"object"}}}},"responses":{"200":{"description":"successful operation","content":{"application\/json":{"schema":{},"examples":{"success":{"summary":"Validation success","value":{"code":"qyg35xy5DWfJQJTl","message":"Success."}},"eligible,locked":{"summary":"Validation failed","value":{"error_message":"Image recoginition failed."}}}}}},"404":{"description":"No voucher locked for user"},"422":{"description":"Validation failed"}},"security":[{"bearer_token":[]}]}}},"components":{"securitySchemes":{"bearer_token":{"type":"http","description":"Enter token","name":"Authorization","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}